Fix kernel ldscript to properly align things
It turns out that 'foo = ALIGNED(.)' ensures the value of 'foo' is aligned,
but this means the section around it could be anything... (we are only
guaranteeing alignment of the foo variable, not the stuff around it)

This can be worked around by explicitely aligning all the sections
and leaving the variables at '.'.

It's a pure coincidence that this didn't trigger earlier.